cmd/go-cache-plugin: rearrange how the HTTP services are set up
The http.ServeMux has some annoying built-in behaviour that is inconvenient for
us here. Specifically, it tries to automatically issue redirects when it cannot
resolve the handler, and that gets in the way of the proxy.

Instead, plumb together a simple direct dispatcher, since we only have a small
number of path options anyway.