lib/revproxy: refine the logic for cacheability
In addition to "immutable", treat an object as effectively cacheable if its
freshness period (max-age) is so long that we never need to worry about
updating it (because the backing store will be long gone).

NPM in particular seems to set expirations to either 5m or 365d, on all the
binary artifacts that one fetches from there.